Docs: create nice 404 pages
Currently, when a user arrives at a 404 page, Read the Docs shows the default
Maze Found page without context about the setuptools project itself.

This extension, `sphinx-notfound-page`, creates a `404.html` page with the same
style as the rest of the documentation pages that Read the Docs will serve
instead of the default Maze Found one.