DOC: Extend Windows rpath error message with likely cause
Extend the error message for attempting to call `runtime_library_dir_option` on Windows with the most likely way I have found to trigger the call, creating an `Extension` that specifies `runtime_library_dirs`.

Hopefully this will give users something to search for in their `setup.py` and a starting point for what to change.  Inspired by pypa/setuptools#3450.