Replicate problems with name discovery and dynamic versions
According to issue 3545 it seems that "name-discovery" happens, even
when the project already explicitly sets it.

This is related to parsing of dynamic versions (via `attr` directive),
which triggers the auto-discovery to obtain the value of `package_dir`.

The value of `package_dir` is used to find the path to the module
in `version = {"attr" = "module_name.attr_name"}`.