Make __file__ an absolute path in setuptools.build_meta
This is a difference between pip's legacy wheel build (or direct
invocations of setup.py) and the PEP 517 build.

While setup.py scripts that rely on this are fragile, it was quite
painful to debug! Since Python 3.4, `__file__` is usually an absolute
path, so this change might result in fewer surprises.