distutils.ccompiler.CCompiler.has_function: Quote #include argument
Arguably, this is a historic wart in the interface, which is why I
subconsciously fixed it in commit 56a5b333b2a8 ("distutils.ccompiler:
Make has_function work with more C99 compilers").  But it's clearly
not a backwards-compatible change, so it's wrong and has to be
reverted.

Fixes pypa/setuptools#3820.