docs: Add a note about historical `wheel` use in `requires`
Add a note explaining that the `wheel` requirement that used to be
historically used in documentation was incorrect.  This also explains
why it can frequently be found in existing projects, and what problems
it causes.  I find myself repeatedly explaining it, so I think having
a single official source would be better.

Incorporates suggestions from Anderson Bravalheri.