Remove dependency on `pip-run`.
The dependency chain of `pip-run` includes `pydantic-core`.
The problem is that newer versions `pydantic-core` are written in Rust
and don't have pre-compiled wheels available for all platforms (e.g.
freebsd). It might be the case it is not even currently possible to
compile it to some platforms due to the lack of availability of the
toolchain (seems to be the problem with cygwin).

See `jaraco/inflect#195` and `pydantic/pydantic-core#773`.