Use the stdlib toml library on sufficiently new python
Although tomli is vendored and always available, it is more clean to
avoid using the backport on python 3.11. This makes it easier to
automatically drop outdated branches in the future, and is a
micro-optimization when user code imports a toml library in the same
process space.

Additionally, setuptools.extern.VendorImporter explicitly supports
removing vendored modules as long as they are globally installed and
available. For distributors that rely on this, importing tomllib first
permits those distributors to avoid packaging tomli on versions of
python that already have tomllib.