Fix PathEntryFinder / MetaPathFinder in editable_wheel
It seems that the import machinery skips PathEntryFinder when trying to
locate nested namespaces, if the `sys.path_hook` item corresponding to
the finder cannot be located in the submodule locations of the parent
namespace.

This means that we should probably always add the PATH_PLACEHOLDER to
the namespace spec.

This PR also add some type hints to the template, because it helped to
debug some type errors.