Avoid setuptools_wheel fixture to create recursively nested build/lib/build/lib/... directories
Based on the test introduced in b4d3e83f0, we can see that when none of
`PRE_BUILT_SETUPTOOLS_SDIST` or `PRE_BUILT_SETUPTOOLS_WHEEL` is set,
the `setuptools_wheel` fixture keeps recursively creating
`build/lib/build/lib/...` directories which slows down the tests and
creates a huge amount of unnecessary files.

This change tries to target that.