Avoid stating in metadata that the project as a whole uses MIT
Despite the code for `setuptools` itself be distributed under MIT,
setuptools includes third-party code with other licenses.

Adding `license = "MIT"` to `pyproject.toml` is problematic under the
lenses of PEP 639 because it may imply that all the files distributed by
setuptools are licensed under MIT which is not the case.

By removing that metadata, setuptools should be "in the clear".